Fixed warnings

+++ src/test/java/se/sics/ace/oscore/group/TestDtlspClientGroupOSCORE.java
@@ -279,13 +279,13 @@ public class TestDtlspClientGroupOSCORE {
          
         Assert.assertEquals(4, joinResponse.size());
         Assert.assertEquals(CBORType.Number, joinResponse.get("kty").getType());
-        Assert.assertEquals((int) 0, joinResponse.get("kty").AsInt32());
+        Assert.assertEquals(0, joinResponse.get("kty").AsInt32());
         Assert.assertEquals(CBORType.Map, joinResponse.get("k").getType());
         Assert.assertEquals(0, joinResponse.get("k").size());
         Assert.assertEquals(CBORType.Number, joinResponse.get("profile").getType());
-        Assert.assertEquals((int) 0, joinResponse.get("profile").AsInt32());
+        Assert.assertEquals(0, joinResponse.get("profile").AsInt32());
         Assert.assertEquals(CBORType.Number, joinResponse.get("exp").getType());
-        Assert.assertEquals((int) 1000000, joinResponse.get("exp").AsInt32());
+        Assert.assertEquals(1000000, joinResponse.get("exp").AsInt32());
         
     }
     
@@ -371,13 +371,13 @@ public class TestDtlspClientGroupOSCORE {
         
         Assert.assertEquals(4, joinResponse.size());
         Assert.assertEquals(CBORType.Number, joinResponse.get("kty").getType());
-        Assert.assertEquals((int) 0, joinResponse.get("kty").AsInt32());
+        Assert.assertEquals(0, joinResponse.get("kty").AsInt32());
         Assert.assertEquals(CBORType.Map, joinResponse.get("k").getType());
         Assert.assertEquals(0, joinResponse.get("k").size());
         Assert.assertEquals(CBORType.Number, joinResponse.get("profile").getType());
-        Assert.assertEquals((int) 0, joinResponse.get("profile").AsInt32());
+        Assert.assertEquals(0, joinResponse.get("profile").AsInt32());
         Assert.assertEquals(CBORType.Number, joinResponse.get("exp").getType());
-        Assert.assertEquals((int) 1000000, joinResponse.get("exp").AsInt32());
+        Assert.assertEquals(1000000, joinResponse.get("exp").AsInt32());
         
     }
     
@@ -499,13 +499,13 @@ public class TestDtlspClientGroupOSCORE {
         
         Assert.assertEquals(4, joinResponse.size());
         Assert.assertEquals(CBORType.Number, joinResponse.get("kty").getType());
-        Assert.assertEquals((int) 0, joinResponse.get("kty").AsInt32());
+        Assert.assertEquals(0, joinResponse.get("kty").AsInt32());
         Assert.assertEquals(CBORType.Map, joinResponse.get("k").getType());
         Assert.assertEquals(0, joinResponse.get("k").size());
         Assert.assertEquals(CBORType.Number, joinResponse.get("profile").getType());
-        Assert.assertEquals((int) 0, joinResponse.get("profile").AsInt32());
+        Assert.assertEquals(0, joinResponse.get("profile").AsInt32());
         Assert.assertEquals(CBORType.Number, joinResponse.get("exp").getType());
-        Assert.assertEquals((int) 1000000, joinResponse.get("exp").AsInt32());
+        Assert.assertEquals(1000000, joinResponse.get("exp").AsInt32());
 
     }
     
@@ -593,13 +593,13 @@ public class TestDtlspClientGroupOSCORE {
         
         Assert.assertEquals(4, joinResponse.size());
         Assert.assertEquals(CBORType.Number, joinResponse.get("kty").getType());
-        Assert.assertEquals((int) 0, joinResponse.get("kty").AsInt32());
+        Assert.assertEquals(0, joinResponse.get("kty").AsInt32());
         Assert.assertEquals(CBORType.Map, joinResponse.get("k").getType());
         Assert.assertEquals(0, joinResponse.get("k").size());
         Assert.assertEquals(CBORType.Number, joinResponse.get("profile").getType());
-        Assert.assertEquals((int) 0, joinResponse.get("profile").AsInt32());
+        Assert.assertEquals(0, joinResponse.get("profile").AsInt32());
         Assert.assertEquals(CBORType.Number, joinResponse.get("exp").getType());
-        Assert.assertEquals((int) 1000000, joinResponse.get("exp").AsInt32());
+        Assert.assertEquals(1000000, joinResponse.get("exp").AsInt32());
 
     }
     
@@ -751,13 +751,13 @@ public class TestDtlspClientGroupOSCORE {
         
         Assert.assertEquals(4, joinResponse.size());
         Assert.assertEquals(CBORType.Number, joinResponse.get("kty").getType());
-        Assert.assertEquals((int) 0, joinResponse.get("kty").AsInt32());
+        Assert.assertEquals(0, joinResponse.get("kty").AsInt32());
         Assert.assertEquals(CBORType.Map, joinResponse.get("k").getType());
         Assert.assertEquals(0, joinResponse.get("k").size());
         Assert.assertEquals(CBORType.Number, joinResponse.get("profile").getType());
-        Assert.assertEquals((int) 0, joinResponse.get("profile").AsInt32());
+        Assert.assertEquals(0, joinResponse.get("profile").AsInt32());
         Assert.assertEquals(CBORType.Number, joinResponse.get("exp").getType());
-        Assert.assertEquals((int) 1000000, joinResponse.get("exp").AsInt32());
+        Assert.assertEquals(1000000, joinResponse.get("exp").AsInt32());
         
     }
     
@@ -842,13 +842,13 @@ public class TestDtlspClientGroupOSCORE {
         
         Assert.assertEquals(4, joinResponse.size());
         Assert.assertEquals(CBORType.Number, joinResponse.get("kty").getType());
-        Assert.assertEquals((int) 0, joinResponse.get("kty").AsInt32());
+        Assert.assertEquals(0, joinResponse.get("kty").AsInt32());
         Assert.assertEquals(CBORType.Map, joinResponse.get("k").getType());
         Assert.assertEquals(0, joinResponse.get("k").size());
         Assert.assertEquals(CBORType.Number, joinResponse.get("profile").getType());
-        Assert.assertEquals((int) 0, joinResponse.get("profile").AsInt32());
+        Assert.assertEquals(0, joinResponse.get("profile").AsInt32());
         Assert.assertEquals(CBORType.Number, joinResponse.get("exp").getType());
-        Assert.assertEquals((int) 1000000, joinResponse.get("exp").AsInt32());
+        Assert.assertEquals(1000000, joinResponse.get("exp").AsInt32());
         
     }
     
@@ -1171,13 +1171,13 @@ public class TestDtlspClientGroupOSCORE {
         
         Assert.assertEquals(4, joinResponse.size());
         Assert.assertEquals(CBORType.Number, joinResponse.get("kty").getType());
-        Assert.assertEquals((int) 0, joinResponse.get("kty").AsInt32());
+        Assert.assertEquals(0, joinResponse.get("kty").AsInt32());
         Assert.assertEquals(CBORType.Map, joinResponse.get("k").getType());
         Assert.assertEquals(0, joinResponse.get("k").size());
         Assert.assertEquals(CBORType.Number, joinResponse.get("profile").getType());
-        Assert.assertEquals((int) 0, joinResponse.get("profile").AsInt32());
+        Assert.assertEquals(0, joinResponse.get("profile").AsInt32());
         Assert.assertEquals(CBORType.Number, joinResponse.get("exp").getType());
-        Assert.assertEquals((int) 1000000, joinResponse.get("exp").AsInt32());
+        Assert.assertEquals(1000000, joinResponse.get("exp").AsInt32());
         
     }
     
@@ -1275,13 +1275,13 @@ public class TestDtlspClientGroupOSCORE {
         
         Assert.assertEquals(4, joinResponse.size());
         Assert.assertEquals(CBORType.Number, joinResponse.get("kty").getType());
-        Assert.assertEquals((int) 0, joinResponse.get("kty").AsInt32());
+        Assert.assertEquals(0, joinResponse.get("kty").AsInt32());
         Assert.assertEquals(CBORType.Map, joinResponse.get("k").getType());
         Assert.assertEquals(0, joinResponse.get("k").size());
         Assert.assertEquals(CBORType.Number, joinResponse.get("profile").getType());
-        Assert.assertEquals((int) 0, joinResponse.get("profile").AsInt32());
+        Assert.assertEquals(0, joinResponse.get("profile").AsInt32());
         Assert.assertEquals(CBORType.Number, joinResponse.get("exp").getType());
-        Assert.assertEquals((int) 1000000, joinResponse.get("exp").AsInt32());
+        Assert.assertEquals(1000000, joinResponse.get("exp").AsInt32());
         
     }
